use joblib to load/save model * use only one file for model + processor * use 2 early stoppings 1) comparable 2) keras-default one
