+++ word2morph/__init__.py
@@ -1,0 +1,1 @@
+from word2morph.api import Word2Morph, load_model
