+++ word2morph/util/callbacks.py
@@ -1,13 +1,12 @@
-from typing import Optional, List
+from typing import List
 
 from keras.callbacks import EarlyStopping
 
 
 class ComparableEarlyStopping(EarlyStopping):
-    def __init__(self, to_compare_values: Optional[List[float]] = None, **kwargs):
+    def __init__(self, to_compare_values: List[float], **kwargs):
         super().__init__(**kwargs)
         self.compare_values = to_compare_values
-        self.compare_wait = 0
         print('Compare values:', self.compare_values)
 
     def on_epoch_end(self, epoch, logs=None):
@@ -15,23 +14,15 @@ class ComparableEarlyStopping(EarlyStopping):
         if current is None:
             return
 
-        ''' Compare the current value to the ones obtained before '''
-        if self.monitor_op(current - self.min_delta, self.best):
-            self.best = current
+        ''' Compare the obtained value to the ones provided in the constructor '''
+        if len(self.compare_values) <= epoch or self.monitor_op(current - self.min_delta, self.compare_values[epoch]):
             self.wait = 0
             if self.restore_best_weights:
                 self.best_weights = self.model.get_weights()
         else:
             self.wait += 1
 
-        ''' Compare the obtained value to the ones provided in the constructor '''
-        if self.compare_values is None or len(self.compare_values) <= epoch or \
-                self.monitor_op(current - self.min_delta, self.compare_values[epoch]):
-            self.compare_wait = 0
-        else:
-            self.compare_wait += 1
-
-        if self.wait + self.compare_wait >= self.patience:
+        if self.wait >= self.patience:
             self.stopped_epoch = epoch
             self.model.stop_training = True
             if self.restore_best_weights:
